cloudscape_trail/clone/requirements.txt
flask==3.0.1
flask-Cors==4.0.0
boto3==1.34.34
pynamodb==6.0.0
pypdf==4.0.1
langchain==0.0.354
langchain-community==0.0.17
langchain-core==0.1.18
langchain-google-genai==0.0.6
python-dotenv==1.0.1
chromadb==0.4.22